fastapi_mongo_project/requirements.txt
fastapi
uvicorn
pymongo
jinja2
passlib
python-multipart
email-validator
pandas
itsdangerous
authlib
python-dotenv
httpx
